Fix: Remove configuration assumptions from priority scheduling test.
Prior to this commit, the test SCHED0005 assumed there were 256 priorities
available. This commit modifies the test to be independent of the number of
priority levels and automatically disables it when there are too few to run
this test. In particular, this makes the test now pass when we reduce the
number of priorities to, e.g., 255.