Support x64 tests
This fixes various tests that were not compatible with x64. The real
changes are to tests that create/map/recycle frames and these have
been rewritten to be clearer and support both 32 and 64bit.

Other tests have minor changes for type compatibility and reserving
additional amounts of memory needed for the changes.