Consolidate the shared test_init_data_t struct
Previously there was a struct that was used to access shared memory between
the driver app and the test child processes. This struct's definition was
duplicated between the two apps.

This patch does an initial fixup for that, by making the shared-mem struct
definition a non-relative include for the driver app, and then symlinking that
to be a non-relative include for the child app as well.