sel4test-tests: init timer if CONFIG_HAVE_TIMER=y
This avoids timer initialisation failures on simulators.
We don't run the timer tests on simulators anyway due to
inaccuarcy of interrupt delivery.

A better solution would be to fall back on different timer
implementations (the current failure is that on x86 our hpet
driver relies on fsb delivery which qemu does not support).