Reduce MAX_LENGTH in ipc tests
Current value of seL4_MsgMaxLength + 1 leads to indexing pass
the end of the 'msg' array in seL4IPCBuffer. This occurs since
FOR_EACH_LENGTH loops while length <= MAX_LENGTH, and then inside
these loops there are loops that iterate while i < length. In final
case where length == seL4_MsgMaxLength + 1 we will get i == seL4_MsgMaxLength,
which if used in an index into an array that is of size seL4_MsgMaxLength
will be 1 past the end of the array.

Aside from being incorrect for the test this is also undefined
behaviour and in some cases the compiler notices this and discards
the top level loop from the test