Avoid multiple initializations of the C library
When spawning helper processes we perform a clone of the current vspace
and then begin running this cloned process from the `_start` entry point.
As the cloned data sections have already been initialized by the C library
this means we effectively initialize the C library again. This can be
especially confusing to any init constructors or other routines that
you expect to run once.

The goal of spawning the helper process is to get it to run a desired
function in the new process. This is controlled by adding additional
arguments to the new process, and having `main` jump to it. Instead
of doing this, this change makes it so that we start directly running
at the desired function, instead of going through `_start` and `main`.
Since we do not pass these extra arguments this change also cleans up
`main` and `helper_thread`