Explicitly check the instruction pointer softbreak test
This acts as a form of documentation-through-source as well.

On x86: the processor executes the INT3 instruction and then triggers the
trap after.

On ARM: the processor delivers the "bkpt" instruction as a fault and not a trap,
so the address of the BKPT instruction itself is delivered to the kernel. This is
consistent with what the ARM manual describes:

ARMv7 arch manual: C3.2:
"BKPT instruction debug events are synchronous. That is, the debug event acts like
an exception that cancels the BKPT instruction. ... A BKPT instruction debug event
generates a Prefetch Abort exception ..."