sel4test-driver: Make different environment types clearer
Changes the 'env' in sel4test-driver to be called driver_env so that it is not confused
with the environment 'env' in the basic tests. To handle this type change we also
explicitly declare implementations that could take an arbitrary environment to take a
'uintptr_t' type that is cast to the known (and hopefully correct) environment type.