Use ZF_LOGF_IF in non test context
These functions, whilst called in the context of a running test, are potentially not
called by the test process, but rather by the driver on behalf of the test. This means
that if the test_eq fails the originating call site, in the test process, will not
know to return the correct result code. As a result, the only thing that can be
done is to generate a fatal error and abort.