cmake: Change PLATFORM to use seL4 platform names
This removes the need for a lot of name mangling.
Add a ARM_HYP flag to specify hyp mode on tk1.
Use AARCH64 and AARCH32 flags used in gcc.cmake instead of having hikey
and hikey64 platforms.
This commit also moves some elfloader settings into a separate if block
that correspond to how we configure our boards in Data61