Remove IPC buffer dependency for regression test.
Due to the nature in which the test thread for the write registers test
is run (without any initialisation), it may not be guaranteed to have
acceess to its IPC buffer.

`seL4_SendWithMRs` is used in place of `seL4_Send` such that no attempt
is made to access the IPC buffer.