sel4test-tests: make PREEMPT_REVOKE more robust
The test expects that the revoke operation will be preempted at least
twice. The old timeout value of half the expected time means that the
second preemption is likely to be missed in some conditions.

Using a timeout of a third the expected time, at least two preemptions
will occur as expected.

Also do not attempt to copy CNodes if the CNode table could not be
allocated.