RISC-V: Use same arch.c files as Arm
These arch.c functions provide implementation for copying serial
hardware capabilities into the test processes. As the RISC-V serial
hardware is similar to Arm (1 4K frame and 1 IRQ) we temporarily reuse
the same arch specific functions. A FIXME has been left to indicate that
the files are symlinked and should be refactored.