sel4test-driver: Invert allocation of serial objs
Previously we allocated objects for serial devices before initialising
the serial. Workarounds were required to allow these objects to be
allocated twice. We reorder this process and capture the objects
allocated by the serial as it is initialised. This removes the need to
manage the structure of objects allocated which could be different per
serial device.