tests,timer: Use time manager instead of ltimer
For some platforms with small timer counter widths, there may be more
than one source of timer interrupts. This could lead to timestamp
overflow events being mistaken for the timeout events. So instead of
going directly to the ltimer to setup timeouts, we use the time manager
to manage the timeouts so that these both events are correctly managed.