Fix SCHED0021 test
SCHED0021 had a bug that meant it was dereferencing beyound the
boundaries of an array. This led to inconsistent behaviour.

The order in which threads are inserted into the scheduling queue is
the opposite to what one would expect: they're entered at the *front*
not the back.  Plus the last thread to run will see the first thread's
updated data and immediately loop --- add a Yield() to prevent that.

Signed-off-by: Peter Chubb <peter.chubb@data61.csiro.au>