mcs: Add tests for invalid SC faults
TIMEOUTFAULT0004: Fault on no donated SC

When a passive thread receives an IPC and the sender has not used
seL4_Call, the sender's SC is not donated and the receiver faults to
indicate that it is running without a SC.

TIMEOUTFAULT0005: Fault on client SC removed

When a passive thread receives an IPC and the sender's SC has been
removed while it is blocked, receiver faults to indicate that it is
running without a SC.

TIMEOUTFAULT0006: Fault on SC removed while running

When a thread with an SC is running and its SC is removed, that thread
faults.

TIMEOUTFAULT0007: Fault on donated SC removed while running

When a passive thread is running on a donated SC and that SC is unbound,
the passive thread faults as it no longer has an SC.

TIMEOUTFAULT0008: Fault on reply without SC return

When a client makes a call that results in a nested call and a monitor
replies to the first call, the client is not returned its SC, as such it
faults.

TIMEOUTFAULT0009: Fault on donation of unconfigured SC

When a passive thread receives an IPC and the sender's SC has been
replaced with an unconfigured SC while it is blocked, receiver faults to
indicate that it is running with an unconfigured SC.

Signed-off-by: Curtis Millar <curtis.millar@data61.csiro.au>