Ensure server prio greater than client for timeout
We must ensure that during the reset of the server in the timeout test
that the client is not able to preempt the server. If it does, it can
perform a send before the server performs a receive. If this occurs, the
server will not be blocked when the test thread removes its scheduling
context and the test will hang indefinitely.

Signed-off-by: Curtis Millar <curtis.millar@data61.csiro.au>