Allow Lazy SC Rebind
It is now allowed to bind a scheduling context to both a notification
object as well as a TCB. When that is the case, the SC will be unbound
from the TCB when the task blocks on that notification, making the task
passive. Just as if the task has been running on a SC bound to the
notification.

Signed-off-by: Indan Zupancic <Indan.Zupancic@mep-info.com>