SCHED0022: do not request unnecessary refills
When this test was introduced in commit beff0afd0bbb6d it requested 1
extra refill in each sched_context_configure call. These extra refills
are not necessary to change the core/affinity of the SC and were
probably an oversight.

The calls succeed when there is enough space in the SchedContext for at
least 4 refills, but fail for 64 bit architectures with the now
(seL4/seL4@bb6a00ffe25) tighter minimum SchedContext size where the
minimum size works out to exactly 2 refills.

Signed-off-by: Gerwin Klein <gerwin.klein@proofcraft.systems>