domain_schedule: pick more useful defaults
The previous schedule lead to 15/16th idle time and had extremely short
(1 tick) domain slices which means a lot of the test time was spent
idle or domain switching instead of running code.

This setup give domain 0 a long-ish slice to make some progress, adds
some variety to the schedule and sets 2 as the shortest period.

Signed-off-by: Gerwin Klein <gerwin.klein@proofcraft.systems>