use config_set only for boolean config values
config_set is true for 1 and false for all other values, so it does
not make sense to use with CONFIG_MAX_NUM_NODES.

The two instances where this happened look like they are trying to
guard CONFIG_MAX_NUM_NODES > 1 by checking whether CONFIG_MAX_NUM_NODES
has any value at all, but doing so incorrectly. CONFIG_MAX_NUM_NODES
has a default value of 1 and should always be set to a number, so we
can just drop the guard. If this assumption about always having a
numbers value is wrong, the code will not compile, which is acceptable.

Signed-off-by: Gerwin Klein <gerwin.klein@proofcraft.systems>