vspace004: allocate a cap for the final pool
This *coincidentally* fails with ret == seL4_RevokeFirst(9) if the
number of pools to allocate before failing is less than the actual
amount, because we don't allocate a capability for the last pool.

In essence, that could be happening even if we had the correct number
of pools, but doesn't because of how the kernel (presumably) checks
that the cap is free after checking the ASID pools are free.

Signed-off-by: julia <git.ts@trainwit.ch>