SCHED_CONTEXT_0014: Use sel4utils thread directly
start_helper would do unnecessary complicated thread startup code
which may contain FPU instructions with some compilers and configs.

Normally this would be harmless (though it would thwart the test's
intention), but if the task runs out of budget before it finishes
the FPU instruction, it will be put on the release queue and make
no forward progress (unclear why).

This is only a problem for tasks with budget equal to MIN_BUDGET,
any time lost because of syscalls or traps will cause the task to
hang.

Signed-off-by: Indan Zupancic <indan@nul.nu>