use LibPlatSupportHaveTimer for sel4test HaveTimer
In libplatsupport, we know which platforms do not support an ltimer
implementation, so rather than duplicating them, we can re-use the
cmake config variable here.

However, this needs to be moved after util_libs_import_libraries()
for the config to appear, but before config_option for it to be
generated in the gen_config.h file.

I have tested this with:

- maaxboard (supports timer, tests enabled)
- zynqmp (supports timer, tests disabled)
- cheshire (no timer support, tests disabled)

Signed-off-by: julia <git.ts@trainwit.ch>