Fix test failures when -fshort-enums is used
When compiled with arm-none-eabi-gcc (13.3.0, and others), the
default ABI has -fshort-enums on by default. This seem to also be
the case for x86 with gcc (GCC) 15.1.1 20250521 (Red Hat 15.1.1-2).

Due to C's arcane integer promotion rules, the enum members such as
SUCCESS or seL4_NoError are of type "int" (generally signed 32 bit),
whereas the size of seL4_Error or test_result_t is 8 bits (a char).

For reference: https://godbolt.org/z/zoKzhqxc7

This produces test output like:

    Error: res (size 1) != SUCCESS (size 4), use of test_eq incorrect

Signed-off-by: julia <git.ts@trainwit.ch>